Separate Position internals
