+++ src/Data/Generics/Product/Positions.hs
@@ -43,18 +43,9 @@ module Data.Generics.Product.Positions
 
 import Data.Generics.Internal.VL.Lens as VL
 import Data.Generics.Internal.Void
-import Data.Generics.Internal.Families
 import Data.Generics.Product.Internal.Positions
-import Data.Generics.Product.Internal.GLens
-import Data.Generics.Internal.Errors
 
-import Data.Kind      (Constraint, Type)
-import Data.Type.Bool (type (&&))
-import GHC.Generics
-import GHC.TypeLits   (type (<=?),  Nat, TypeError, ErrorMessage(..))
-import Data.Generics.Internal.Profunctor.Lens as P
-import Data.Generics.Internal.Profunctor.Iso as P
-import Data.Coerce
+import GHC.TypeLits   (Nat)
 
 -- $setup
 -- == /Running example:/
@@ -127,43 +118,13 @@ getPosition s = s ^. position' @i
 setPosition :: forall i s a. HasPosition' i s a => a -> s -> s
 setPosition = VL.set (position' @i)
 
-instance
-  ( Generic s
-  , ErrorUnless i s (0 <? i && i <=? Size (Rep s))
-  , cs ~ CRep s
-  , Coercible (Rep s) cs
-  , GLens' (HasTotalPositionPSym i) cs a
-  , Defined (Rep s)
-    (NoGeneric s '[ 'Text "arising from a generic lens focusing on the field at"
-                  , 'Text "position " ':<>: QuoteType i ':<>: 'Text " of type " ':<>: QuoteType a
-                    ':<>: 'Text " in " ':<>: QuoteType s
-                  ])
-    (() :: Constraint)
-  ) => HasPosition' i s a where
-  position' f s = VL.ravel (repIso . coerced @cs @cs . glens @(HasTotalPositionPSym i)) f s
+instance Context' i s a => HasPosition' i s a where
+  position' f s = VL.ravel (derived' @i) f s
   {-# INLINE position' #-}
 
-instance  -- see Note [Changing type parameters]
-  ( ErrorUnless i s (0 <? i && i <=? Size (Rep s))
-  , HasTotalPositionP i (CRep s) ~ 'Just a
-  , HasTotalPositionP i (CRep t) ~ 'Just b
-  , HasTotalPositionP i (CRep (Indexed s)) ~ 'Just a'
-  , HasTotalPositionP i (CRep (Indexed t)) ~ 'Just b'
-  , t ~ Infer s a' b
-  , s ~ Infer t b' a
-  , HasPosition0 i s t a b
-  ) => HasPosition i s t a b where
-
+instance (Context i s t a b , HasPosition0 i s t a b) => HasPosition i s t a b where
   position = position0 @i
   {-# INLINE position #-}
-
--- We wouldn't need the universal 'x' here if we could express above that
--- forall x. Coercible (cs x) (Rep s x), but this requires quantified
--- constraints
-coerced :: forall s t s' t' x a b. (Coercible t t', Coercible s s')
-        => P.ALens a b (s x) (t x) -> P.ALens a b (s' x) (t' x)
-coerced = coerce
-{-# INLINE coerced #-}
 
 -- | See Note [Uncluttering type signatures]
 -- >>> :t position
@@ -172,48 +133,13 @@ coerced = coerce
 instance {-# OVERLAPPING #-} HasPosition f (Void1 a) (Void1 b) a b where
   position = undefined
 
-instance
-  ( ErrorUnless i s (0 <? i && i <=? Size (Rep s))
-  , UnifyHead s t
-  , UnifyHead t s
-  , Coercible (CRep s) (Rep s)
-  , Coercible (CRep t) (Rep t)
-  , HasPosition0 i s t a b
-  ) => HasPosition_ i s t a b where
-
+instance (Context_ i s t a b, HasPosition0 i s t a b) => HasPosition_ i s t a b where
   position_ = position0 @i
   {-# INLINE position_ #-}
 
 instance {-# OVERLAPPING #-} HasPosition_ f (Void1 a) (Void1 b) a b where
   position_ = undefined
 
-instance
-  ( Generic s
-  , Generic t
-  , GLens (HasTotalPositionPSym i) (CRep s) (CRep t) a b
-  , Coercible (CRep s) (Rep s)
-  , Coercible (CRep t) (Rep t)
-  , Defined (Rep s)
-    (NoGeneric s '[ 'Text "arising from a generic lens focusing on the field at"
-                  , 'Text "position " ':<>: QuoteType i ':<>: 'Text " of type " ':<>: QuoteType a
-                    ':<>: 'Text " in " ':<>: QuoteType s
-                  ])
-    (() :: Constraint)
-  ) => HasPosition0 i s t a b where
-  position0 = VL.ravel (repIso . coerced @(CRep s) @(CRep t) . glens @(HasTotalPositionPSym i))
+instance Context0 i s t a b => HasPosition0 i s t a b where
+  position0 f s = VL.ravel (derived0 @i) f s
   {-# INLINE position0 #-}
-
-type family ErrorUnless (i :: Nat) (s :: Type) (hasP :: Bool) :: Constraint where
-  ErrorUnless i s 'False
-    = TypeError
-        (     'Text "The type "
-        ':<>: 'ShowType s
-        ':<>: 'Text " does not contain a field at position "
-        ':<>: 'ShowType i
-        )
-
-  ErrorUnless _ _ 'True
-    = ()
-
-data HasTotalPositionPSym  :: Nat -> (TyFun (Type -> Type) (Maybe Type))
-type instance Eval (HasTotalPositionPSym t) tt = HasTotalPositionP t tt
